added queries page
